Move the mqtt fast path guard beside its siblings
